@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,15 +27,16 @@ dependencies:
   flutter_bloc: ^7.0.1
 
   # Firebase
-  # firebase_core: ^1.4.0
+  firebase_core: ^1.4.0
   # firebase_database: ^7.1.2
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  # IOS Icons
   cupertino_icons: ^1.0.2
-  smooth_page_indicator: ^0.2.3
-  step_progress_indicator: ^1.0.0
+
+  # Page View dots indicator
   dots_indicator: ^2.0.0
+
+  # Percent circle
   percent_indicator: ^3.0.1
 
 dev_dependencies:
@@ -54,6 +55,7 @@ flutter:
   assets:
     - assets/icons/
     - assets/images/onboarding/
+    - assets/images/auth/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
