@@ -8,5 +8,5 @@ class PageChangedEvent extends OnboardingEvent {}
 class PageSwipedEvent extends OnboardingEvent {
   final int index;
 
-  PageSwipedEvent({@required this.index});
+  PageSwipedEvent({required this.index});
 }
