@@ -8,5 +8,9 @@ class OnboardingInitial extends OnboardingState {}
 class PageChangedState extends OnboardingState {
   final int counter;
 
-  PageChangedState({@required this.counter});
+  PageChangedState({
+    required this.counter,
+  });
 }
+
+class NextScreenState extends OnboardingState {}
