@@ -1,8 +1,15 @@
 import 'package:flutter/material.dart';
 
 class ColorConstants {
-  static const kTextColor = Color(0xFF1F2022);
-  static const kPrimaryColor = Color(0xFF6358E1);
+  static const textColor = Color(0xFF1F2022);
+  static const primaryColor = Color(0xFF6358E1);
 
-  static const inactiveDot = Color(0xFF6358E1);
+  static const textBlack = Color(0xFF1F2022);
+  static const white = Color(0xFFFFFFFF);
+  static const grey = Color(0xFFB6BDC6);
+
+  static const textFieldBackground = Color(0xFFFBFCFF);
+  static const textFieldBorder = Color(0xFFB9BBC5);
+
+  static const disabledColor = Color(0xFFE1E1E5);
 }
