@@ -1,6 +1,7 @@
 import 'package:fitness_flutter/screens/onboarding/bloc/onboarding_bloc.dart';
 import 'package:fitness_flutter/screens/onboarding/widget/onboarding_content.dart';
 import 'package:fitness_flutter/core/config/size_config.dart';
+import 'package:fitness_flutter/screens/sign_up/page/sign_up_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -15,10 +16,19 @@ class OnboardingPage extends StatelessWidget {
 
   BlocProvider<OnboardingBloc> _buildBody(BuildContext context) {
     return BlocProvider<OnboardingBloc>(
-      create: (BuildContext context) => OnboardingBloc(),
-      child: BlocBuilder<OnboardingBloc, OnboardingState>(
-        buildWhen: (_, OnboardingState currState) =>
-            currState is OnboardingInitial,
+      create: (BuildContext context) => onboardingBloc,
+      child: BlocConsumer<OnboardingBloc, OnboardingState>(
+        listenWhen: (_, currState) => currState is NextScreenState,
+        listener: (context, state) {
+          Navigator.of(context).pushReplacement(
+            MaterialPageRoute(
+              builder: (BuildContext context) {
+                return SignUpPage();
+              },
+            ),
+          );
+        },
+        buildWhen: (_, currState) => currState is OnboardingInitial,
         builder: (context, state) {
           return OnboardingContent();
         },
